+++ pyproject.toml
@@ -32,6 +32,7 @@ black = "^21.12b0"
 flake8 = "^4"
 isort = "^5.9.2"
 pyright = "^0.0.13"
+coverage = "^6.2"
 
 [tool.poetry.extras]
 json = ["orjson"]
